Add `className` to sortable properties in Astro
When using Astro with React you’d also want to sort `className` in addition to `class` and `class:list`. While we might be able to do some work to figure out if the element the attirubte is on is a React component it seems unncessary and I think it’s reasonable to have this enabled all the time.