Introduce `tailwindStylesheet` option to replace `tailwindEntryPoint` (#330)
* Refactor

* Introduce `tailwindStylesheet` option to replace `tailwindEntryPoint`

The old option will still work but the new one is much better named

* Allow `tailwindConfig` to be used as a stylesheet option

This is a compatability feature. People previously used the `tailwindConfig` option to specify the location of their config and naturally expect it to work with a CSS file in v4. The preferred option is `tailwindStylesheet` but making this work is rather simple so it should “just work” for improved DX

* Warn when using outdated config options

* Update readme

* Bump Tailwind version in tests

* Update option in tests

* Update README.md

Co-authored-by: Jonathan Reinink <jonathan@reinink.ca>

* Update README.md

Co-authored-by: Jonathan Reinink <jonathan@reinink.ca>

* Update README.md

Co-authored-by: Jonathan Reinink <jonathan@reinink.ca>

* Tweak title

---------

Co-authored-by: Jonathan Reinink <jonathan@reinink.ca>